+++ packages/react-vapor/src/components/limit/Limit.tsx
@@ -1,7 +1,8 @@
 import classNames from 'classnames';
 import * as VaporSVG from 'coveo-styleguide';
 import * as React from 'react';
-import {connect} from 'react-redux';
+import {useDispatch, useSelector} from 'react-redux';
+
 import {IReactVaporState} from '../../ReactVaporState';
 import {IDispatch} from '../../utils/ReduxUtils';
 import {InputConnected} from '../input';
@@ -19,123 +20,77 @@ export interface LimitOwnProps {
     isLimitEditable?: boolean;
     isLimitTheGoalToReach?: boolean;
     isHistoryIncluded?: boolean;
+    onHistoryIconClick?: () => void;
 }
 
-const mapStateToProps = (state: IReactVaporState, {id, limit}: LimitOwnProps) => {
-    const ownLimit: number = limit ?? 100;
-    const currentLimit: number = +InputSelectors.getValue(state, {id}) || ownLimit;
-    return {
-        currentLimit,
-    };
+export const Limit: React.FunctionComponent<LimitOwnProps> = (props) => {
+    const {id, limit, className} = props;
+    const {currentLimit} = useSelector((state: IReactVaporState) => ({
+        currentLimit: +InputSelectors.getValue(state, {id}) || limit,
+    }));
+
+    return (
+        <div className={classNames('limit-box mb2', className)}>
+            <div className="limit-box-main p2 pb1">
+                <HeaderDivision {...props} />
+                <ContentDivision {...props} currentLimit={currentLimit} />
+            </div>
+            <ProgressBar usage={props.usage} isLimitTheGoalToReach={props.isLimitTheGoalToReach} limit={currentLimit} />
+        </div>
+    );
 };
 
-const mapDispatchToProps = (dispatch: IDispatch, {id}: LimitOwnProps) => ({
-    onChangeLimit: (limitValue: string) => {
-        dispatch(changeInputValue(id, limitValue, true));
-    },
-});
-
-export interface LimitProps
-    extends LimitOwnProps,
-        ReturnType<typeof mapDispatchToProps>,
-        ReturnType<typeof mapStateToProps> {}
-
-type ContentProps = Pick<LimitProps, 'id' | 'usage' | 'limit' | 'isLimitEditable' | 'limitLabel' | 'onChangeLimit'>;
-type UsageContentProps = Pick<ContentProps, 'usage'>;
-type LimitContentProps = Pick<
-    ContentProps,
-    'id' | 'limit' | 'usage' | 'isLimitEditable' | 'limitLabel' | 'onChangeLimit'
->;
-type HeaderProps = Pick<LimitProps, 'limitLabel' | 'isHistoryIncluded'>;
-type HistoryIconProps = Pick<HeaderProps, 'isHistoryIncluded'>;
-type ProgressBarProps = Pick<LimitProps, 'usage' | 'isLimitTheGoalToReach' | 'limit'>;
-
-const LimitDisconnect: React.FunctionComponent<LimitProps> = ({
-    id,
-    title,
-    limitLabel = 'Limit',
-    usage,
-    limit = 100,
-    currentLimit,
-    isLimitEditable = false,
-    isLimitTheGoalToReach = false,
-    isHistoryIncluded = false,
-    onChangeLimit,
-    className,
-}) => (
-    <div className={classNames('limit-box mb2', className)}>
-        <div className="limit-box-main p2 pb1">
-            <HeaderDivision limitLabel={title} isHistoryIncluded={isHistoryIncluded} />
-            <ContentDivision
-                id={id}
-                usage={usage}
-                limit={limit}
-                isLimitEditable={isLimitEditable}
-                limitLabel={limitLabel}
-                onChangeLimit={onChangeLimit}
-            />
-        </div>
-        <ProgressBar usage={usage} isLimitTheGoalToReach={isLimitTheGoalToReach} limit={currentLimit} />
-    </div>
-);
-
-const HeaderDivision: React.FunctionComponent<HeaderProps> = ({limitLabel, isHistoryIncluded}) => (
+const HeaderDivision: React.FunctionComponent<LimitOwnProps> = ({title, isHistoryIncluded, onHistoryIconClick}) => (
     <div className="flex space-between">
-        <label className="form-control-label"> {limitLabel}</label>
-        <HistoryIcon isHistoryIncluded={isHistoryIncluded} />
+        <label className="form-control-label"> {title}</label>
+        <HistoryIcon isHistoryIncluded={isHistoryIncluded} onHistoryIconClick={onHistoryIconClick} />
     </div>
 );
 
-const HistoryIcon: React.FunctionComponent<HistoryIconProps> = ({isHistoryIncluded}) =>
-    isHistoryIncluded && (
-        <span className="icon mod-lg limit-history-button">
+const HistoryIcon: React.FunctionComponent<Omit<LimitOwnProps, 'title'>> = ({isHistoryIncluded, onHistoryIconClick}) =>
+    isHistoryIncluded ? (
+        <span className="icon limit-history-button documentation-link" onClick={onHistoryIconClick}>
             <Svg svgName={VaporSVG.svg.menuAnalytics.name} />
         </span>
-    );
+    ) : null;
 
-const ContentDivision: React.FunctionComponent<ContentProps> = ({
+const ContentDivision: React.FunctionComponent<Omit<LimitOwnProps, 'title'> & {currentLimit: number}> = ({
     id,
     usage,
     limit,
     isLimitEditable,
-    limitLabel,
-    onChangeLimit,
+    limitLabel = 'Limit',
+    currentLimit,
 }) => (
     <div className="limit-box-numbers pt1 flex">
         <UsageDivision usage={usage} />
         <LimitDivision
             id={id}
-            onChangeLimit={onChangeLimit}
             usage={usage}
             limit={limit}
-            isLimitEditable={isLimitEditable}
-            limitLabel={limitLabel}
+            isLimitEditable={!!currentLimit && isLimitEditable}
+            limitLabel={currentLimit ? limitLabel : ''}
         />
     </div>
 );
 
-const UsageDivision: React.FunctionComponent<UsageContentProps> = ({usage}) => {
-    const isUsageRequired: boolean = !!usage;
-    return (
-        isUsageRequired && (
-            <div className="limit-box-usage">
-                <label className="form-control-label">Usage</label>
-                <span className="limit-box-usage-value">{usage}</span>
-            </div>
-        )
-    );
-};
+const UsageDivision: React.FunctionComponent<Omit<LimitOwnProps, 'title'>> = ({usage}) => (
+    <div className="limit-box-usage">
+        <label className="form-control-label">Usage</label>
+        <span className="limit-box-usage-value">{usage ?? 0}</span>
+    </div>
+);
 
-const LimitDivision: React.FunctionComponent<LimitContentProps> = ({
+const LimitDivision: React.FunctionComponent<Omit<LimitOwnProps, 'title'>> = ({
     id,
     limit,
     usage,
     isLimitEditable,
     limitLabel,
-    onChangeLimit,
 }) => {
-    const limitValueString: string = limit.toString();
+    const limitValueString: string = limit?.toString();
     const minLimitValue: number = usage ?? 0;
+    const dispatch: IDispatch = useDispatch();
     return isLimitEditable ? (
         <InputConnected
             id={id}
@@ -144,29 +99,26 @@ const LimitDivision: React.FunctionComponent<LimitContentProps> = ({
             defaultValue={limitValueString}
             min={minLimitValue}
             classes="limit-box-limit form-group input-field validate"
-            onChange={(limitValue: string) => onChangeLimit(limitValue)}
+            onChange={(limitValue: string) => dispatch(changeInputValue(id, limitValue, true))}
         />
     ) : (
         <div className="limit-box-limit">
-            <label className="form-control-label">{limitLabel}</label>
+            <label className="form-control-label">{limit && limitLabel}</label>
             <span className="limit-box-limit-value">{limit}</span>
         </div>
     );
 };
 
-const ProgressBar: React.FunctionComponent<ProgressBarProps> = ({usage, isLimitTheGoalToReach, limit}) => {
-    const isUsageRequired: boolean = !!usage;
+const ProgressBar: React.FunctionComponent<Omit<LimitOwnProps, 'title'>> = ({usage, isLimitTheGoalToReach, limit}) => {
     const progressClass: string = getProgressBarClass(usage, limit);
     const progressClasses = classNames('limit-box-bar', progressClass, {
         'mod-green': isLimitTheGoalToReach,
     });
 
     return (
-        isUsageRequired && (
-            <div className="limit-box-footer">
-                <div className={progressClasses}></div>
-            </div>
-        )
+        <div className={classNames('limit-box-footer', {'no-limit': !limit})}>
+            <div className={[usage, limit].every((value) => Number.isInteger(value)) ? progressClasses : ''}></div>
+        </div>
     );
 };
 
@@ -174,5 +126,3 @@ const getProgressBarClass = (usageValue: number, limitValue: number): string => 
     const progress: number = Math.round((usageValue / limitValue) * 100);
     return progress <= 100 ? `progress-${progress.toString()}` : `progress-100`;
 };
-
-export const Limit = connect(mapStateToProps, mapDispatchToProps)(LimitDisconnect);